Split constraint_system.rs into a folder module
The single file had grown to ~2600 lines holding a dozen unrelated types
plus ~1350 lines of tests. Turn it into a folder module with one file per
concept, colocating each type's tests with the code they exercise.

The public API is unchanged: mod.rs re-exports every submodule, so all
external `binius_core::constraint_system::{...}` paths keep resolving.

Pure code motion. The only non-mechanical touches:
- `ValueVecLayout::is_padding` / `is_committed_oob` become `pub(super)`
  because `system.rs::validate` now calls them across a file boundary.
- The three `include_bytes!` reference-binary paths gain one `../`.
- Four doc-only intra-doc links that now cross a module boundary get
  qualified targets so they still resolve.
- `test_word_serialization_round_trip` moves to word.rs, its true home.
