+++ crates/core/src/constraint_system/mod.rs
@@ -1,0 +1,20 @@
+// Copyright 2025 Irreducible Inc.
+//! Constraint system and related definitions.
+
+mod constraint;
+mod layout;
+mod proof;
+mod shift;
+mod system;
+mod value_index;
+mod value_vec;
+mod values_data;
+
+pub use constraint::*;
+pub use layout::*;
+pub use proof::*;
+pub use shift::*;
+pub use system::*;
+pub use value_index::*;
+pub use value_vec::*;
+pub use values_data::*;
